requirements.txt
streamlit==1.40.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.2
scipy==1.14.0
mplsoccer==1.4.0
pillow==10.3.0
pathlib==1.0.1
python-dotenv==1.0.1
img2pdf==0.4.4
fastparquet==2024.2.0
pyarrow==15.0.0
mysql-connector-python==8.3.0
duckdb>=0.9.0
thefuzz
python-Levenshtein
polars==0.20.9
reportlab==4.0.9
plotly==5.18.0
scikit-learn==1.2.2
